Thread-safe barrier recreation
If a thread is done with its share of the insertion work it
needs to bail out of the pool. Therefore, new barriers have to
be created to reduce the number of waiting threads.
Previously, a thread could still be stuck in the conditional
variable during the release step while another already reentered
the method and deleted the barrier and thus the variable.

Had to introduce a read-write lock to resolve this issue.
Before entering the flush phase each thread acquires a read lock
and releases it at the end of the method. When a thread enters
the flushing method and detects that the barrier needs to be changed
it tries to get unique write permissions before actually deleting
the barrier. This introduces only light overhead as
write locks are only needed if any thread is done (i.e. for N
threads at most N times).