Reduced write lock time for subtree creation
If a buffer is transformed into a subtree it holds a write lock
for an extended time period. Instead of holding a write lock
an therefore blocking all further reading threads on the current
node only an upgradable locks is acquired which ensures that
the write privilege is available.

The subtree creation routine returns the root node of the new
subtree. Only when this root node is inserted into the parent
of the buffer the parent actually needs to be locked.

Introduced a new flag that indicates whether the insertion process
is done so the flushing does not start to early.