Started to parallelize the subtree creation
When a buffer is flushed one thread currently blocks all further
access to the node. Thus, this is a high contention point.

Started to implement a helping routine where new threads can
enter the subtree creation by sharing the work between
threads.

As a first step had to place all data in the class so all threads
have access to it. In a next step the threads get different roles
during the flush:
- the first thread is the coordinator and is responsible for
incorporating newly arrived threads
- all further threads join the subtree creation, once they are done
they wait to be released in order to continue from the current node
- the coordinator finally adds the root of the subnode to the
current node and than releases all helper threads before continuing