Extended and refactored parallel buffer flush
Refactored the steps needed to flush a buffer.
Especially, separated the creation of nodes and the
insertion of suffixes / subnodes into them.

Fixed several deadlock problems of threads that got
stuck in the conditional variables.

Before a thread inserts an entry into a node it first of all
needs to varify if there is a previous node that is going to
insert the current node:
	a...	b...	c...
a		x	x

b		()	x

c

x - maximum LCP value so needs to be inserted in current round
() - thread checks whether a previous row has the same x
--> in this case the thread responsible for a has to handle b and c
instead of b handling only c!