External locks for nodes
Instead of placing the read-write locks inside a node and thus
introducing a bigger memory footprint which is not needed after
the creation of the tree the locks are now placed in an external
table.

This table provides three means of handling locks:
1) find lock -> a thread that does not hold a lock on a given node
yet can get a lock, i.e. reuse an existing lock or find a new one
2) use lock -> a thread that already found a lock can use it
arbitrarily often
3) free lock -> an acquired lock can thus be freed

The table consumes much less memory than providing a lock per node.
Also cache lines are invalidated for the node if contents of the
lock change which might be a bottleneck for high numbers of threads.
It stores:
2 references (node pointers) per thread, 2 indices per thread,
2 locks per thread and 1 counter per lock.
That means that locks can be reused for different nodes.
The design goal of the approach was to avoid locking within this
table. Instead a lock-free approach based on CAS operations was
pursued.

The basic idea of the data structure is to map a reference to
an index that indicates which lock is used. When finding a lock all
of the references of other nodes need to be traversed to find the index
that holds the lock which is currently in use for a given node.
The indices can hold to more values than the actual lock index:
- (SEARCHING) meaning that the current thread is searching for the
current reference
- (SEARCH_SURRENDER) meanng that the current thread is searching but
waits for the result of a higher priority thread
- (INVALID - verification only) to indicate that the value is invalid

Attention: The current solution needs to count the number of reading
threads for a given index. Therefore, it uses the 3 least significant
bits which are not used for the actual pointer. A count of the current
readers is kept in these 3-bit. That means a deadlock can occure for
more than 9 threads running in parallel. TO circumvent this the
lock tables can be assigned to ranges of most significant reference bits.
These can thereafter be shifted up to increase the available bits for
the counter.