Bulk: compare to previous entry and skip nodes
Previously each entry to be inserted invokes a series of locks
until the potentially deep node is reached. That means that
especially upper nodes are very congested in terms of readers.
To avoid this congestion and to make use of the ordered structure

During insertion always keep the last inserted entry.
When inserting do the following:
1. Compare the current entry to the previous entry and check
in which node they differ. Start the insertion at the parent
node.
2. Get a read lock an the lowest common node's parent
and validate if it was already deleted (if so need to
linearly backup to the next higher node that was not
deleted).
3. While inserting:
3.1. If a subnode is found insert it with the current index
and in the map.
3.2. If a node needs to be enlarged but there was not parent
node this is an indicator for a too low start. Need to
force the start at the parent.

In order for the node skipping to work nodes must not be
deleted while other threads can potentially access them.
Therefore, introduced a list that stores references of nodes
to be deleted. In the synchronized phase each thread performs
all the deletions. During the parallel phase a deleted node
is marked with a flag.

Added a more efficient routine that compares two multi-dim
bit sequences by making use of the property that both entries
are aligned and have the same number of bits.

Remark:
For random data more than the last entry can be hold in a ring
buffer. A new entry is than compared to all of the previous entries
and uses the insertion details of the one with the longest
common prefix.