Rethought the synchronization of entry buffers
The previous idea for synchronizing the entry buffer spilling
required to many synchronization mechanisms which would
probably be a big overhead for buffers of low capacities.

Instead bigger buffers should be used. While inserting keep
track of the number of different threads. If there is more than
a single thread inserting the parallel flushing routine should
be chosen. Otherwise stop inserting at few entries and flush in
a sequential manner.