Atomic LHC, AHC and suffix storages
Implemented a version of the LHC node - Parallel LHC = PLHC -
 which supports atomic updates.
In detail, it supports lock-free insertion and updates of any
suffix / subnode / reference assuming that updates to node contents
cannot accure at the same time as node resizing. This is ensured
by acquiring a write lock before changing the node (e.g. enlarge)
The node works roughly as follows:
1) The first part of the node works exactly as the normal LHC node
and is not supposed to be extended during the lifetime of the node.
That means only updates can be performed in this sorted part.
2) The second part is an unsorted array that can be searched and
updated (insertion, update) in a synchronized manner.
3) When this node needs to be enlarged the unsorted part is sorted
and merged with the previously sorted first part. (Implemented
in the PLHCIterator)

Any already inserted address can be updated to a spinlock which
forces other ready threads to spin on the reference value until
it changes. Adapted the different insertion cases to first
insert a spinlock at the potentially conflicting address and
then execute any required changes on that edge. A spinlock can
be exchanged for a valid reference by the thread that introduced
it. With this mechanism the locking can be done more fine grained
resulting in less restarts due to write lock misses.
Example: (assuming write locks are held)
a) Inserting a suffix and enlarging the node: insert a spinlock
in the parent, acquire a write lock on the child, enlarge
the child, insert into the child, exchange the spinlock with
the reference of the enlarged child
b) Split suffix: insert spinlock in node, create the subnode,
exchange the spinlock for the new subnode

In order to assure enough suffix storage space the maximum
storage required for the current node is always allocated up
front and is never shrunk down afterwards.

Also had to implement an atomic version of the suffix storage.
It uses a small occupied map to validate the used contents. A bit
is set for any valid suffix block range in use. When a new suffix
should be inserted any bit is tried to be set in a structured
manner using the CAS operation. Once this succeded the thread
uniquely owns the suffix block range corresponding to the set bit.

Additionaly, extended the AHC node to enable parallel operations
which is conceptually as any HC address uniquely defines the
corresponding array position.

Introduced a try-fail mechanism for all insert and update
operations on a node. Added corresponding linearInsert calls
which ensure a structured insertion order. Thereby, when
copying (P)LHC nodes the sorted part does not need to be searched
but can easily be appended to. Thus, reducing the copy time from
O(N log N) to O(N).

Adjusting the size of a node is no longer handled within the nodes
but was placed in the NodeTypeUtil instead. This utility class can
now differentiate between atomic and non-atomic nodes