Fixed locking related bugs (mostly assertions)
Since the routines such as suffix insertion can now be called
in parallel several of the assertions are not always true any more.
Had to remove all that where not related to unaccessable objects.

Fixed some problems related to removed entries. Has to set the
removed flag before the node is unlocked. Thereby several assertions
had to be removed.