Bug fixes for concurrency issues, visitors, delete
Enabled tracking of parent nodes for PLHC. The nodes need to be
converted back to LHC nodes once the building is done. Therefore,
keeping track of the parent into which the transformed LHC node
needs to be inserted.

Preliminary implementation of recursive deletion with PLHC nodes.
Otherwise the memory consumption was too high for test series.

Implemented visitors for PLHC to keep track of the properties of
this node type.

Bug fixes for atomic parallel insertion:
- in case a suffix is added it might happen that the node needs
to be extended first which can only be permitted if the parent
edge is locked
- recheck if the buffer still needs to be flushed once the
upgradable permission is granted
- previously did not delete suffix storage of an outdated node