Fixed 2 major issues in synchronization
1) When a buffer was entered the address contents were already
predetermined. Thus, if the buffer was flushed in the mean time
and handed back to the pool the thread would still try to insert
into the buffer. Therefore, does not use an upgradable lock any
more (which also prevented further changes on the node.) Instead
the edge locking is used to introduce a spinlock for all further
readers. That means the correct buffer address needs to be
CASed which is not possible if the buffer was flushed because in
that case the contents are replaced with a subnode.

2) Compared HC values to references. Stupid.