Faster concurrent buffer pool flush
Instead of iterating over the entire buffer sequentially to
find the buffers that are in use a boolean flag was added to
each buffer that is maintained by the pool during allocate and
deallocate. Thus, each thread can check locally whether a buffer
needs to be flushed or not.

Also reduced the information output of histograms in various
visitors. Only rows with a value > 0 are printed.