Replaced boost::shared_mutex with pthread::rwlock
Because of the error prune implementation of the boost library
and many struggles repaced the locks with an implementation
of pthreads. These only provide read and write locks and not
the intermediate upgradable state as shared_mutex did. Also
state changes are not guaranteed to be atomic. Had to reimplement
the locking details. Especially, had to introduce an update
counter to validate a synchronized state between call to
pthreads. Only used blocking locks at the moment so there is
only a need for a restart if a node that is supposed to be locked
was deleted / changed in the mean time.